### 2019-07-30 #332 @SteveViss
Submission: rmangal
Submitting Author: Steve Vissault (@SteveViss)
Repository: https://github.com/mangal-wg/rmangal
Version submitted: 1.9.1
Version accepted: TBD

---

```
Package: rmangal
Title: Mangal Client
Version: 1.9.1
    person(given = "Steve", family = "Vissault", comment = c(ORCID = "0000-0002-0866-4376"),
      email = "steve.vissault@usherbrooke.ca", role = c("aut", "cre")),
    person(given = "Kevin", family = "Cazelles", comment = c(ORCID = "0000-0001-6619-9874"),
        email = "kcazelle@uoguelph.ca", role = c("aut","ctb")),
    person(given = "Gabriel", family = "Bergeron",
      email = "gabriel.bergeron3@usherbrooke.ca", role = c("aut","ctb")),
    person(given = "Benjamin", family = "Mercier",
      email = "Benjamin.B.Mercier@usherbrooke.ca", role = c("aut","ctb")),
    person(given = "Clément", family = "Violet",
      email = "Clement.Violet@etudiant.univ-brest.fr", role = c("aut","ctb")),
    person(given = "Dominique", family = "Gravel",
      email = "dominique.gravel@usherbrooke.ca", role = c("aut")),
    person(given = "Timothée", family = "Poisot",
      email = "timothee.poisot@umontreal.ca", role = c("aut"))
      )
Description: An interface to the Mangal database - a collection of ecological networks
<http://poisotlab.biol.umontreal.ca/#/>.
This package includes functions to work with the Mangal RESTFul API methods
(<https://mangal-wg.github.io/mangal-api/#api-documentation>).
    jsonlite (>= 1.5),
    igraph,
    sf
    mapview,
    taxize,
    USAboundaries
```

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences).

The `rmangal` package provides all the functions to retrieve ecological networks stored in [Mangal interactions database v2](https://mangal.io/). The `rmangal` package makes easier to integrate the data acquisition step of ecological networks into a reproducible workflow.

Please note any areas you are unsure of: `reproducibility`

This package targets mostly ecologists interested in the structure of interactions among species,
individus and populations at different scales. By giving access to a large collection of networks around the globe, Mangal opens the path to study ecological networks properties (connectance, degree, trophic structure etc.) at global scale in a more reproducible way.

The only package that could have a potential overlap is `rglobi`. `rmangal` serves a different purpose by giving access at ecological networks (already published) in a standardized matter; whether GloBI is aggregating pairwise interactions among species and do not keep track of the study context (e.g. sampling date and location). Mangal 1.0 was a data provider of the GloBi infrastructure, and we plan to do the same with Mangal 2.0.

- Any other questions or issues we should be aware of?

No

---

Accepted presubmission by @noamross in #318

---

**Yes**, we are currently writing an abstract describing this package. In this paper, we would also like to introduce the **[julia client](https://github.com/PoisotLab/Mangal.jl)** in order to streamline references to Mangal clients. That being said, we are not quite sure about what does this mean on your end, we believe that this implies that the Julia client will need to be reviewed independently.

----
Best,
Steve Vissault
on the behalf of all authors / contributors (@KevCaz, @gabrielbouleau, @BenMerSci, @clementviolet, @tpoisot).
