### 2019-07-30 4559c0f
Rebuild doc (#82)
* Update doc / mispelling and typo

* Add verbose

* Rebuild website

* rebuild doc

* Update print message

* Rebuild docs
