+++ R/search_interactions.R
@@ -21,12 +21,13 @@
 #' Alternatively, a named list can be used to look for an exact match in a specific field.
 #' In this case, the name of the list should match one of the field names of the database table.
 #' For the `interaction` table, those are:
-# - attr_id: identifier of a specific attribute
-# - direction: edge direction ("directed", "undirected" or "unknown")
-# - network_id: Mangal network identifier
-# - node_from: node id which the interaction end to
-# - node_to: node to which the interaction end to
-# - type: use argument `type` instead.
+#' - id: unique identifier of the interaction
+#' - attr_id: identifier of a specific attribute
+#' - direction: edge direction ("directed", "undirected" or "unknown")
+#' - network_id: Mangal network identifier
+#' - node_from: node id which the interaction end to
+#' - node_to: node to which the interaction end to
+#' - type: use argument `type` instead.
 #' Note that for lists with more than one element, only the first element is used, the others are ignored.
 #' Examples covering custom queries are provided below.
 
@@ -34,11 +35,11 @@
 #' <https://mangal-wg.github.io/mangal-api/#interactions>
 #'
 #' @examples
-#' df_inter <- search_interactions(type = "competition")
+#' df_inter <- search_interactions(type = "competition", verbose = FALSE)
 #' # Get all networks containing competition
 #' competition_networks <- get_collection(df_inter)
-#' df_net_926 <- search_interactions(list(network_id = 926))
-#' df_net_unknown <- search_interactions(list(attr_id = 100))
+#' df_net_926 <- search_interactions(list(network_id = 926), verbose = FALSE)
+#' df_net_unknown <- search_interactions(list(attr_id = 100), verbose = FALSE)
 #' @export
 
 search_interactions <- function(query, type = NULL, expand_node = FALSE,
@@ -51,7 +52,7 @@ search_interactions <- function(query, type = NULL, expand_node = FALSE,
       query <- list(type = type)
     } else {
       query <- handle_query(query,
-        c("attr_id", "direction", "network_id", "node_from", "node_to", "type"))
+        c("id" ,"attr_id", "direction", "network_id", "node_from", "node_to", "type"))
     }
 
     # Get interactions based on the type
