+++ R/search_networks.R
@@ -17,10 +17,10 @@
 #' partial match was found are returned.
 #' Alternatively, a named list can be used to look for an exact match in a specific field.
 #' In this case, the name of the list should match one of the field names of the database table. For the `networks` table, those are:
-#' * all_interactions: identifier of a specific attribute;
-#' * id : Mangal identifier of the network;
-#' * dataset_id: the identifier of the dataset;
-#' * public: network publicly available;
+#' - id: unique identifier of the network
+#' - all_interactions: false interaction can be considered as real false interaction;
+#' - dataset_id: the identifier of the dataset;
+#' - public: network publicly available;
 #' Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.
 #'
 #' @references
@@ -46,8 +46,7 @@ search_networks <- function(query, verbose = TRUE, ...) {
 
   if ("sf" %in% class(query))
     return(search_networks_sf(query, verbose, ...))
-  query <- handle_query(query, c("public", "all_interactions", "dataset_id",
-   "id"))
+  query <- handle_query(query, c("id", "public", "all_interactions", "dataset_id"))
 
   networks <- resp_to_spatial(get_gen(endpoints()$network, query = query,
     verbose = verbose, ...)$body)
