+++ R/combine_mgNetworks.R
@@ -9,8 +9,8 @@
 #' An object of class `mgNetworksCollection`
 #'
 #' @examples
-#' mg_19 <- get_collection(search_networks(list(dataset_id = 19)))
-#' mg_lagoon <- get_collection(search_datasets(query='lagoon%'))
+#' mg_19 <- get_collection(search_networks(list(dataset_id = 19)), verbose = FALSE)
+#' mg_lagoon <- get_collection(search_datasets(query='lagoon%'), verbose = FALSE)
 #' combine_mgNetworks(mg_19, mg_lagoon)
 #'
 #' @export
