### 2019-08-04 #332 @noamross
Thanks for your submission, @SteveViss and co-authors!  I am now seeking reviewers.

Regarding JOSS, from our end this simply means that we pass our reviews to the JOSS editors for them to use in place of new reviews.  As you have multiple clients and would like to include them in a single submission to JOSS,  I suggest you submit to JOSS and request they review the Julia client, and we will pass forward the reviews for the R client.  This could wait until this review is finished or you could submit pre-emptively so that the Julia review starts, as well (as you may want to address review comments jointly should they affect common/upstream logic).

### Editor checks:

---

#### Editor comments

`goodpractice::gp()` tests pass with flying colors.

---

Reviewers:
Due date:
