### 2019-08-09 #332 @tpoisot
@noamross we will submit to JOSS asap -- the Julia code is really tight (under 700 sloc), so review should be fast.
