### 2019-08-22 #332 @thomasp85
Hi All

Please find my review below. All in all a very good job with this package 🙂

best
Thomas

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

The rmangal package provides a clean and consistent API for interacting with the
Mangal database. It is in general very well done and signals high quality to the
user. Still, a few issues need to be addressed before approval but all of these
are pretty minor and can easily be handled by the development team. My specific
comments are below:

#### Website
The website is good, and as it is hosted under the main service it wraps it
gives credibility to the whole package. One issue that could be adressed is the
reference index where a bit of care could be used to group functions and give
each group a short description. This will help new users a lot when navigating
the website.

#### Dependencies
The hard dependency on sf could be lifted as it doesn't appear to be used
outside of a single functionality and sf has some very strong system
dependencies. I suggest moving sf to suggest and testing for its availability
when needed along with an error message about installing it if missing.

#### Check
R CMD check reports a range of examples with run time > 5sec and a total runtime
of examples of 53 sec. This is likely to be flagged by CRAN during submission
and it might as well be fixed beforehand:

```
✔  checking examples (53.4s)
   Examples with CPU or elapsed time > 5s
                        user system elapsed
   get_collection      4.812  0.069  22.876
   combine_mgNetworks  1.129  0.056   6.831
   search_interactions 0.718  0.010   6.440
   search_datasets     0.369  0.008   5.359
```

The README is a tad short. It does include installation instructions and some
very useful links, but there is no motivating text for why this package may be
of interest and what it tries to solve (besides being an API wrapper for a
service). It should at least contain descriptions of why one might want to use
the mangal database and an example of a network retrieval to show the strength
of the API

#### Vignette
This is a very good introduction to the package and shows of all important
features in an easy to understand way. I would personally include a short
section on visualising the networks, and maybe drag in `tidygraph` and `ggraph`
(disclaimer: Author of both) for a more familiar interface to network analysis
and visualisation than igraph (at least to the tidyverse crowd). I will not
insist on this, though as package preferences are subjective

A few of the smaller functions lack examples, such a `avail_type()` and
`clear_cache_rmangal()`. These are both functions without arguments so it is not
clear that any example would add to the understanding, but in the interest of
completeness I'll just mention this.
Most of the documentations have very short, sub-title like descriptions, with
most of the text relegated to the *Details* section. This generally result in
documentation entries where one have to read quite a bit before the function is
justified and the experience could be improved by having a proper paragraph in
the *Description* section.
Cross referencing could be improved. All functions that work with mgNetwork or
mgNetworkCollection objects should link to or describe how one might obtain such
an object. The same is true for functions working with mgSearch* objects. The
key `search_*()` could also benefit from better crosslinking by using the
'@family` keyword from roxygen2.
There are documentation entries for all methods to generics (pointing to the
same doc). This makes the documentation index a bit messy so if this could be
solved it would add to the quality.
Small issues: combine_mgNetworks has a stray `#'` in the `...` argument

The igraph integration is nice. With a very little modification you can make it
work directly with tidygraph and ggraph.

1. Either provide an explicit `as.igraph` method for `mgNetwork` object, which
   could simply be an alias for `mg_to_igraph`. Tidygraph and ggraph will look
   for such a method if no other exist and use that.
2. Provide an `as_tbl_graph` method directly. This can again just be a wrapper
   around `mg_to_igraph`, e.g.
   ```r
   as_tbl_graph.mgNetwork <- function(x, ...) {
     as_tbl_graph(mg_to_igraph(x))
   }
   ```
   The obvious benefit of 1) is that you wouldn't need to list tidygraph as a
   dependency, while conversely the benefit of 2) is that you make support
   explicit

#### Unit testing
The package has a very good unit test coverage on all aspects of the code. The
only area lacking is testing error handling as unit tests only tend to cover
whether functions behave as expected with proper input. Tests for whether
functions fails as expected and fails gracefully in the face of problems or bad
input would be icing on the cake.
