### 2019-08-26 #332 @arw36
Here is my review. Apologies coming in right at the deadline, first ROpenSci review...

Best,
Anna

Not be minor 'utility' packages, such as 'thin' API clients.

---

`rmangal` provides access and navigation of the Mangal database of ecological networks. The ease of this package combined with the valuable information stored in Mangal is an exciting and timely contribution to ecological research, and really expands the scale of between-network comparisons possible. I am particularly impressed with the integration of various taxonomic authorities, citation databases, and the addition of spatial information to published networks that will greatly enhance the flexibility of the collated dataset. I also really like showing the % progress of each dataset as it loads. Overall, a promising package with good coverage and usability. Here are some things that can improve the package.

#### Minor Comments

- It would be easy to prepopulate standard metrics. For the network datatable this would include # of nodes and interactions, potentially a modularity factor, and the interactions database with degree and eigen.centrality.
- Unclear what the `df_net_unknown <- search_interactions(list(attr_id = 100), verbose = FALSE)` example is meant to accomplish in the search_interactions help file. This returned an empty dataframe for me.
- May want to add `get_collection()` to the mg_insect example for `search_networks` documentation
- Not sure on policy for package dependencies when including help page Examples, but may want to suggest `USAboundaries` if using in example functions for `search_networks`and `taxize` if showing integration.
- Suggest you add California border to the plot(networks_in_area) example
- I would like to see more metadata on variables including definitions for `attribute.id` and associated columns (name, description, unit) as well as `user_id`.
- For the date in the edges table, it is unclear when this is extracted as a network-level trait, is an trait specific to each edge, or is NA. For instance, for edges of network_id = 19 it appears to be extrapolated from the network trait, despite the dataset_id (1) giving a range of dates (1897-09 to 1899-07).
- Each data table starts with an `id` column, it may be worth clarifying: `network_id`, `edge_id`, `node_id`, e.t.c. in the main table. This will allow easy linking across tables, as ids are referenced by this in other tables (e.g. the `network_id` column in the nodes table).
- I suggest you make all your functions plural (currently `search_reference` is the only singular one.
- I suggest you rename the author column in the reference table to be `first_author`
- I was able to retrieve complete data tables for network, dataset, node, interactions using `search_*(“”)`. This did not work for the reference table. Ease of full database download would be nice, perhaps with warnings about download times and file sizes.
- I think there should be greater curation/spell checks done in the interaction table, method column. What does biblio mean? Standardize capitalization.
- I find the categorization of interaction types to be non-standard. For instance, I would standardize null, NA, and none of interactions$attribute.unit to be the same (or clarify NA vs. none).
- I think there is some confusion between edges/interactions. I wish the use of interaction was standardized in output text.
- Currently, you can only `search_taxonomy` by the identified taxon rank. It would be really nice for broader rank taxon searches at the order or phylum level to retrieve all entries, even those that have a smaller scale taxonomy. For instance, as of now search_taxonomy(“Arthropoda”) returns no entries, despite there being lots of arthropods in the datasets.
- To conform with ROpenSci packaging guidelines, you will need to add brief demonstration usage, a description of how similar packages (i.e. `rglobi`) relate to this one, and citation information.
- In general, the ReadMe is short. Keep in mind this ReadMe may be first point of entry, so could be fleshed out a bit more with material you already have in vignettes/website.
- Do you expect users to upload networks to this database like `rglobi`? May be worth saying something about contributing new data or identifying new papers.

#### Typos
- In the get_network_by_id.R file *identifiera* should be *identifiers*.
- In the rmangal.Rmd file. *embbeded* should be *embedded*.
- In the “Get started with rmangal” page, section Data Structures, nodes - should be “individual”. This also comes up the `search_nodes` help page and rmangal.Rmd.
- Line 197 of rmangal.Rmd should be *lagoon* instead of *laggoon*.
- Line 379 of rmangal.Rmd should be *the* instead of *tha*
- Line 3 of NEWS.md *submited* should be *submitted.*
- Line 3 of search_reference.Rd. *key wor* should be *keyword*
