### 2019-08-28 75dee5a
Reduce timelaps `get_collection()` in example
