### 2019-08-28 ebc4cf0
Reduce timelaps in combine_mgNetworks() example
