### 2019-08-28 621dcb8
Reduce timelapse in search_* examples
