### 2019-08-28 f31417c
Fix long timelaps in examples
