+++ R/search_networks.R
@@ -36,7 +36,7 @@
 #' plot(networks_in_area)
 #' }
 #' # Network 5013
-#' net_5013 <- search_networks(query = list(id = "5013"))
+#' net_5013 <- search_networks(query = list(id = 5013))
 #' # Network(s) of dataset id = 19
 #' mg_19 <- search_networks(list(dataset_id = 19))
 #'
