+++ man/get_collection.Rd
@@ -61,10 +61,8 @@ function or a set of Mangal unique network identifiers.
 }}
 
 \examples{
-mg_2 <- get_collection(c(1035:1036))
-mg_insect <- get_collection(search_networks(query='insect\%'))
-mg_lagoon <- get_collection(search_datasets(query='lagoon\%'))
-
+mg_2 <- get_collection(c(1076:1077), verbose = FALSE)
+mg_anemone <- get_collection(search_networks(query='anemone\%'), verbose = FALSE)
 }
 \seealso{
 \code{\link[=search_datasets]{search_datasets()}}, \code{\link[=search_interactions]{search_interactions()}}, \code{\link[=search_networks]{search_networks()}},
