+++ man/search_datasets.Rd
@@ -47,7 +47,6 @@ class(all_datasets)
 }
 # Search with keyword
 mg_lagoon <- search_datasets(query = 'lagoon', verbose = FALSE)
-mg_2011 <- search_datasets(query = '2011', verbose = FALSE)
 # Search with a custom query (specific column)
 mg_kemp <- search_datasets(query = list(name = 'kemp_1977'), verbose = FALSE)
 mg_16 <- search_datasets(query = list(ref_id = 16), verbose = FALSE)
