+++ man/search_interactions.Rd
@@ -51,7 +51,6 @@ df_inter <- search_interactions(type = "competition", verbose = FALSE)
 # Get all networks containing competition
 competition_networks <- get_collection(df_inter, verbose = FALSE)
 df_net_926 <- search_interactions(list(network_id = 926), verbose = FALSE)
-df_net_unknown <- search_interactions(list(attr_id = 100), verbose = FALSE)
 }
 \references{
 \url{https://mangal-wg.github.io/mangal-api/#interactions}
