### 2019-08-28 mangal-wg/rmangal#83 @KevCaz
Remove sf from dependencies list
As suggested by the reviewer https://github.com/ropensci/software-review/issues/332#issuecomment-518006617 let's remove sf from the dependencies:

- [x] remove `search_networks_sf() `
- [x] remove associated tests
- [x] make an example in the vignette how to do this kind of sorting.
