### 2019-08-28 mangal-wg/rmangal#83 @SteveViss
Related to d5a00d5
