+++ R/search_nodes.R
@@ -1,6 +1,11 @@
-#' Query the nodes
+#' Query nodes
 #'
-#' Search for networks by querying the nodes database table.
+#' Search for networks by querying the nodes table.
+#'
+#' If the `query` is a character string, then all character columns in the table
+#' are searched and the entries for which at least one
+#' partial match was found are returned.
+#' Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
 #'
 #' @param query either a character string including a single keyword or a list containing a custom query (see details section below).
 #' Note that if an empty character string is passed, then all datasets available are returned.
@@ -8,11 +13,7 @@
 #' @param ... further arguments to be passed to [httr::GET()].
 #'
 #' @details
-#' If `query` is a character string, then all fields of the database table
-#' including character strings are searched and entries for which at least one
-#' partial match was found are returned.
-#' Alternatively, a named list can be used to look for an exact match in a specific field.
-#' In this case, the name of the list should match one of the field names of the table.
+#' Names of the list should match one of the column names within the table. 
 #' For the `networks` table, those are:
 #' - id: unique identifier of the nodes
 #' - original_name: taxonomic name as in the original publication;
