+++ R/search_datasets.R
@@ -1,6 +1,11 @@
-#' Query the datasets
+#' Query datasets
 #'
 #' Identify relevant datasets using a keyword or a custom query.
+#'
+#' If the `query` is a character string, then all character columns in the table
+#' are searched and the entries for which at least one
+#' partial match was found are returned.
+#' Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
 #'
 #' @param query either a character string including a single keyword or a list containing a custom query (see details section below).
 #' Note that if an empty character string is passed, then all datasets available are returned.
@@ -12,11 +17,7 @@
 #' including all datasets corresponding to the query. For each dataset entry, #' the networks and the original reference are attached.
 #'
 #' @details
-#' If `query` is a character string, then all fields of the database table
-#' including character strings are searched and entries for which at least one
-#' partial match was found are returned.
-#' Alternatively, a named list can be used to look for an exact match in a specific field.
-#' In this case, the name of the list should match one of the field names of the database table.
+#' Names of the list should match one of the column names within the table. 
 #' For the `dataset` table, those are:
 #' - id: unique identifier of the dataset
 #' - name: name of the dataset;
