+++ R/search_reference.R
@@ -1,6 +1,11 @@
-#' Query the references
+#' Query references
 #'
 #' Search for a specific reference using a key wor or a Digital Object Identifier (DOI).
+#'
+#' If the `query` is a character string, then all character columns in the table
+#' are searched and the entries for which at least one
+#' partial match was found are returned.
+#' Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
 #'
 #' @param query either a character string including a single keyword or a list containing a custom query (see details section below).
 #' Note that if an empty character string is passed, then all datasets available are returned.
@@ -12,11 +17,7 @@
 #' An object of class `mgSearchReference`, which is a list that includes a wide range of details associated to the reference, including all datasets and networks related to the publication that are included in Mangal database.
 #'
 #' @details
-#' If `query` is a character string, then all fields of the database table
-#' including character strings are searched and entries for which at least one
-#' partial match was found are returned.
-#' Alternatively, a named list can be used to look for an exact match in a specific field.
-#' In this case, the name of the list should match one of the field names of the database table.
+#' Names of the list should match one of the column names within the table. 
 #' For the `reference` table, those are:
 #' - id: unique identifier of the reference
 #' - author: first author;
