+++ man/search_networks.Rd
@@ -3,7 +3,7 @@
 \name{search_networks}
 \alias{search_networks}
 \alias{search_networks_sf}
-\title{Query the networks}
+\title{Query networks}
 \usage{
 search_networks(query, verbose = TRUE, ...)
 
@@ -26,11 +26,13 @@ An object of class \code{mgSearchNetworks}, which is a \code{data.frame} object 
 Search over all networks using a keyword, a custom query or a spatial object
 }
 \details{
-If \code{query} is a character string, then all fields of the database table
-including character strings are searched and entries for which at least one
+If the \code{query} is a character string, then all character columns in the table
+are searched and the entries for which at least one
 partial match was found are returned.
-Alternatively, a named list can be used to look for an exact match in a specific field.
-In this case, the name of the list should match one of the field names of the database table. For the \code{networks} table, those are:
+Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
+
+Names of the list should match one of the column names within the table.
+For the \code{networks} table, those are:
 \itemize{
 \item id: unique identifier of the network
 \item all_interactions: false interaction can be considered as real false interaction;

[…]
