### 2019-08-28 mangal-wg/rmangal#83 @SteveViss
I choose to leave the function `search_network_sf()`. If a `sf` object is passed to the function and the package is not installed, then the user will receive an error recommending to install it.
