+++ man/search_nodes.Rd
@@ -20,13 +20,12 @@ All networks in which taxa are involved are also attached to the \code{data.fram
 }
 \description{
 Search for networks by querying the nodes table.
-}
-\details{
 If the \code{query} is a character string, then all character columns in the table
 are searched and the entries for which at least one
 partial match was found are returned.
 Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
-
+}
+\details{
 Names of the list should match one of the column names within the table.
 For the \code{networks} table, those are:
 \itemize{
