+++ man/search_interactions.Rd
@@ -25,14 +25,12 @@ All networks in which interactions are involved are also attached to the \code{d
 }
 \description{
 Search for specific interactions using a keyword or a specific type of
-interactions (e.g. mutualism).
-}
-\details{
-If the \code{query} is a character string, then all character columns in the table
+interactions (e.g. mutualism). If the \code{query} is a character string, then all character columns in the table
 are searched and the entries for which at least one
 partial match was found are returned.
 Alternatively, a named list can be used to look for an exact match in a specific column (see Details section)
-
+}
+\details{
 Names of the list should match one of the column names within the table.
 For the \code{interaction} table, those are:
 \itemize{
