### 2019-08-29 a98475b
Make all functions plurial
