+++ man/search_networks.Rd
@@ -31,14 +31,15 @@ Alternatively, a named list can be used to look for an exact match in a specific
 }
 \details{
 Names of the list should match one of the column names within the table.
-For the \code{networks} table, those are:
+For the \code{networks} table, those are
 \itemize{
 \item id: unique identifier of the network
-\item all_interactions: false interaction can be considered as real false interaction;
-\item dataset_id: the identifier of the dataset;
-\item public: network publicly available;
-Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.
+\item all_interactions: false interaction can be considered as real false interaction
+\item dataset_id: the identifier of the dataset
+\item public: network publicly available
 }
+
+Note that for lists with more than one element, only the first element is used, the others are ignored. An example is provided below.
 }
 \section{Functions}{
 \itemize{
@@ -63,5 +64,5 @@ mg_19 <- search_networks(list(dataset_id = 19))
 
 }
 \references{
-\url{https://mangal-wg.github.io/mangal-api/#networks}
+More metadata on \url{https://mangal-wg.github.io/mangal-api/#networks}
 }
