### 2019-08-31 cd1430d
Changes edges to interactions
