### 2019-08-31 1fa3568
Adapt mgNetwork names
