+++ R/search_interactions.R
@@ -32,7 +32,7 @@
 #' Examples covering custom queries are provided below.
 
 #' @references
-#' More metadata on <https://mangal-wg.github.io/mangal-api/#interactions>
+#' Metadata available at <https://mangal-wg.github.io/mangal-api/#interactions>
 #'
 #' @examples
 #' df_inter <- search_interactions(type = "competition", verbose = FALSE)
