### 2019-09-01 1d47483
mg_to_igraph is now igraph() :art:
