+++ tests/testthat/helper-rmangal.R
@@ -1,2 +1,2 @@
-cl_df <- c("tbl_df", "tbl",  "data.frame")
+cl_df <- c("tbl_df", "tbl", "data.frame")
 nm_co <-  c("network", "nodes", "edges", "dataset", "reference")
