+++ vignettes/rmangal.R
@@ -1,0 +1,16 @@
+## ----setup, include = FALSE----------------------------------------------
+knitr::opts_chunk$set(
+  message = FALSE,
+  collapse = TRUE,
+  comment = "#>"
+)
+# library(sf)
+
+## ----echo = FALSE--------------------------------------------------------
+knitr::include_graphics("figs/data_structure.svg")
+
+## ----load_packages-------------------------------------------------------
+library(rmangal)
+library(magrittr) # for the pip %>%
+library(tibble) # to use tibbles, enhanced data frames
+
