+++ man/as.igraph.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/as.igraph.R
+\name{as.igraph.mgNetwork}
+\alias{as.igraph.mgNetwork}
+\alias{as.igraph.mgNetworksCollection}
+\title{Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects.}
+\usage{
+\method{as.igraph}{mgNetwork}(x, ...)
+
+\method{as.igraph}{mgNetworksCollection}(x, ...)
+}
+\arguments{
+\item{x}{either a \code{mgNetworksCollection} or a \code{mgNetwork} object.}
+
+\item{...}{currently ignored.}
+}
+\value{
+An object of class \code{igraph} for a \code{mgNetwork} object and a list of
+\code{igraph} objects for \code{mgNetworksCollection}.
+}
+\description{
+Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects.
+}
+\section{Methods (by class)}{
+\itemize{
+\item \code{mgNetwork}: Convert \code{mgNetwork} objects to \code{igraph} objects.
+
+\item \code{mgNetworksCollection}: Convert \code{mgNetworksCollection} objects to a list of \code{igraph} objects.
+}}
+
