+++ man/mg_to_igraph.Rd
@@ -1,43 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/mg_to_igraph.R
-\name{mg_to_igraph}
-\alias{mg_to_igraph}
-\alias{mg_to_igraph.mgNetwork}
-\alias{mg_to_igraph.mgNetworksCollection}
-\title{Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects.}
-\usage{
-mg_to_igraph(x, ...)
-
-\method{mg_to_igraph}{mgNetwork}(x, ...)
-
-\method{mg_to_igraph}{mgNetworksCollection}(x, ...)
-}
-\arguments{
-\item{x}{either a \code{mgNetworksCollection} or a \code{mgNetwork} object.}
-
-\item{...}{currently ignored.}
-}
-\value{
-An object of class \code{igraph} for a \code{mgNetwork} object and a list of
-\code{igraph} objects for \code{mgNetworksCollection}.
-}
-\description{
-Coerce \code{mgNetworksCollection} or \code{mgNetwork} objects to \code{igraph} objects.
-}
-\section{Methods (by class)}{
-\itemize{
-\item \code{mgNetwork}: Convert\code{mgNetwork} objects to \code{igraph} objects.
-
-\item \code{mgNetworksCollection}: Convert \code{mgNetworksCollection} objects to \code{igraph} objects.
-}}
-
-\examples{
-insects_networks <- get_collection(search_networks(query='insect\%'))
-# Apply mg_to_igraph on one specific network
-insects_network <- insects_networks[[1]]
-ig_network <- mg_to_igraph(insects_network)
-# Apply mg_to_igraph on networks collection
-ig_coll_networks <- mg_to_igraph(insects_networks)
-# Plot igraph object with vertex label
-plot(ig_network, vertex.label = insects_network$nodes$taxonomy.name)
-}
