+++ man/reexports.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/as.igraph.R
+\docType{import}
+\name{reexports}
+\alias{reexports}
+\alias{as.igraph}
+\title{Objects exported from other packages}
+\keyword{internal}
+\description{
+These objects are imported from other packages. Follow the links
+below to see their documentation.
+
+\describe{
+  \item{igraph}{\code{\link[igraph]{as.igraph}}}
+}}
+
