+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(as.igraph,mgNetwork)
+S3method(as.igraph,mgNetworksCollection)
 S3method(get_citation,mgNetwork)
 S3method(get_citation,mgNetworksCollection)
 S3method(get_collection,default)
@@ -9,10 +11,9 @@ S3method(get_collection,mgSearchNetworks)
 S3method(get_collection,mgSearchNodes)
 S3method(get_collection,mgSearchReference)
 S3method(get_collection,mgSearchTaxonomy)
-S3method(mg_to_igraph,mgNetwork)
-S3method(mg_to_igraph,mgNetworksCollection)
 S3method(print,mgNetwork)
 S3method(print,mgNetworksCollection)
+export(as.igraph)
 export(avail_type)
 export(clear_cache_rmangal)
 export(combine_mgNetworks)
@@ -20,10 +21,10 @@ export(get_citation)
 export(get_collection)
 export(get_network_by_id)
 export(get_network_by_id_indiv)
-export(mg_to_igraph)
 export(search_datasets)
 export(search_interactions)
 export(search_networks)
 export(search_nodes)
 export(search_reference)
 export(search_taxonomy)
+importFrom(igraph,as.igraph)

[…]
