### 2019-09-02 cc297c8
review search_references() :art:
