### 2019-09-03 7872b50
add summary methods :sparkles:
