+++ NAMESPACE
@@ -13,6 +13,8 @@ S3method(get_collection,mgSearchReferences)
 S3method(get_collection,mgSearchTaxonomy)
 S3method(print,mgNetwork)
 S3method(print,mgNetworksCollection)
+S3method(summary,mgNetwork)
+S3method(summary,mgNetworksCollection)
 export(as.igraph)
 export(avail_type)
 export(clear_cache_rmangal)
