+++ .travis.yml
@@ -33,6 +33,16 @@ matrix:
             - libproj-dev
             - libgeos-dev
             - libgdal-dev
+      r_github_packages:
+        - r-lib/covr
+      before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
+      deploy:
+        provider: script
+        script: Rscript -e 'pkgdown::deploy_site_github()'
+        skip_cleanup: true
+      after_success:
+        - Rscript -e 'covr::codecov()'
+
 notifications:
   slack:
     rooms: insileco:c6N7g4s36dNHpAxrI2QdattF
@@ -46,14 +56,4 @@ notifications:
     on_success: never # change or always
     on_failure: never # change or always
 
-r_github_packages:
-  - r-lib/covr
 
-before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
-deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github()'
-  skip_cleanup: true
-
-after_success:
-  - Rscript -e 'covr::codecov()'
