### 2019-09-03 d6f3eeb
Fix fig issue in docs
