+++ vignettes/rmangal.Rmd
@@ -66,7 +66,7 @@ networks added and the API entirely rewritten. Consequently, [the first release 
 <div class = "row">
 <div class = "col-md-6">
 ```{R echo = FALSE}
-knitr::include_graphics("figs/data_structure.svg")
+knitr::include_graphics("man/figures/data_structure.svg")
 ```
 </div>
 <div class = "col-md-6">
