### 2019-09-03 b1c0d83
Improve README and docs structure
