+++ _pkgdown.yml
@@ -15,3 +15,28 @@ authors:
     href: https://github.com/BenMerSci
   Timothée Poisot:
     href: https://github.com/tpoisot
+
+reference:
+- title: Explore database content
+  desc: Explore the content of each datatables and return `mgSearch*` objects
+  contents:
+  - '`search_datasets`'
+  - '`search_interactions`'
+  - '`search_networks`'
+  - '`search_nodes`'
+  - '`search_references`'
+  - '`search_taxonomy`'
+- title: Retrieve networks
+  desc: Get networks (structured as `mgNetwork` or `mgNetworksCollection`) from `search_*` functions 
+  contents:
+  - '`get_collection`'
+  - '`get_network_by_id`'
+- title: Analyze networks
+  desc: Analyze or manipulate `mgNetwork` or `mgNetworksCollection` 
+  contents:
+  - '`as.igraph.mgNetwork`'
+  - '`combine_mgNetworks`'
+- title: Cache
+  desc: Manipulate `rmangal` cache
+  contents:
+  - '`clear_cache_rmangal`'
