+++ NEWS.md
@@ -1,4 +1,11 @@
+# rmangal 1.9.1.9000
+
+* Revisions see https://github.com/ropensci/software-review/issues/332
+* add summary method
+* `mg_to_igraph` is now `as.igraph()`
+* vignette now includes examples to use tigygraph and ggraph
+
 # rmangal 1.9.1
 
-* Version submited to ROpenSci
+* Version submitted to ROpenSci
 * Added a `NEWS.md` file to track changes to the package.
