+++ vignettes/rmangal.Rmd
@@ -55,8 +55,6 @@ networks added and the API entirely rewritten. Consequently, [the first release 
 **rmangal v2.0** in this vignette.
 
 
-
-
 ## Data structure
 
 
@@ -65,7 +63,7 @@ networks added and the API entirely rewritten. Consequently, [the first release 
 
 <div class = "row">
 <div class = "col-md-6">
-<img src="https://raw.githubusercontent.com/mangal-wg/rmangal/gh-pages/reference/figures/data_structure.png" alt="Mangal data structure"/>
+<img src="man/figures/data_structure.png" alt="Mangal data structure"/>
 </div>
 <div class = "col-md-6">
 
