+++ vignettes/rmangal.Rmd
@@ -63,7 +63,7 @@ networks added and the API entirely rewritten. Consequently, [the first release 
 
 <div class = "row">
 <div class = "col-md-6">
-<img src="man/figures/data_structure.png" alt="Mangal data structure"/>
+<img src="https://raw.githubusercontent.com/mangal-wg/rmangal/gh-pages/reference/figures/data_structure.png" alt="Mangal data structure"/>
 </div>
 <div class = "col-md-6">
 
