+++ .travis.yml
@@ -33,6 +33,7 @@ matrix:
             - libproj-dev
             - libgeos-dev
             - libgdal-dev
+            - libcairo2-dev
       r_github_packages:
         - r-lib/covr
       before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
