+++ .travis.yml
@@ -22,6 +22,10 @@ matrix:
     - os: osx
       r: release
       env: NOT_CRAN=true
+      addons:
+        homebrew:
+          packages:
+          - cairo
     - os: linux
       dist: xenial
       r: devel
@@ -33,7 +37,6 @@ matrix:
             - libproj-dev
             - libgeos-dev
             - libgdal-dev
-            - libcairo2-dev
       r_github_packages:
         - r-lib/covr
       before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
