### 2019-09-03 cdb2898
review resp_to_spatial() :art:
