### 2019-09-03 0264bec
review resp_to_spatial() :art:
