### 2019-09-04 0e8c9ee
search_networks_sf() is now exported
