+++ tests/testthat/test-search_networks.R
@@ -4,7 +4,7 @@ library(sf)
 library(USAboundaries)
 res1 <- search_networks("lagoon")
 area <- us_states(state="california")
-res2 <- search_networks(area)
+res2 <- search_networks_sf(area)
 res2b <- search_networks_sf(area)
 resw <- search_networks("does not exist")
 ress <- search_networks(query="insect%")
