### 2019-09-04 c21f83b
add stop_if_missing_sf() :sparkles:
