### 2019-09-04 4076266
introduce as_sf argument :gear:
