+++ NEWS.md
@@ -1,11 +1,12 @@
 # rmangal 1.9.1.9000
 
-* Revisions see https://github.com/ropensci/software-review/issues/332
-* add summary method
-* `mg_to_igraph` is now `as.igraph()`
-* vignette now includes examples to use tigygraph and ggraph
-* pkgdown website is now deployed by Travis CI
-* query with spatial (`sf`) objects are handle in `query_networks_sf()` that is now exported. 
+* Revisions see https://github.com/ropensci/software-review/issues/332;
+* add summary method;
+* `mg_to_igraph` is now `as.igraph()`;
+* vignette now includes examples to use tigygraph and ggraph;
+* pkgdown website is now deployed by Travis CI;
+* `geom` column has been removed from `mgSearchInteractions` objects;
+* query with spatial (`sf`) objects are handle in `query_networks_sf()` that is now exported.
 
 # rmangal 1.9.1
 
