### 2019-09-04 72979e7
illustrate how to use sf with rmangal :pencil:
