+++ vignettes/rmangal.Rmd
@@ -126,12 +126,12 @@ are under consideration for future major releases.
 
 
 In order to efficiently retrieve networks from the database, **rmangal**
-includes 6 search functions querying the 5 tables described above as well as a table dedicated to the taxonomy backbone.
+includes 7 search functions querying the 5 tables described above as well as a table dedicated to the taxonomy backbone.
 
 
 1. `search_references()`: search in the reference table, for instance the user can look for a specific `doi`;
 2. `search_datasets()`: search among datasets using a keyword;
-3. `search_networks()`: search networks based on a keyword or a geographical area;
+3. `search_networks()` and `search_networks_sf()`: search networks based on a keyword or a geographical area;
 4. `search_interactions()`: list all networks containing a specific interaction type;
 5. `search_nodes()`: identify nodes based on nodes information;
 6. `search_taxonomy()`: identify nodes based on taxonomic names and unique identifiers.
@@ -148,10 +148,11 @@ networks <- search_*() %>% get_collection()
 ```
 
 
-Note that if there is only one network to be retrieved, `get_collection()` returns a `mgNetwork`
-object, otherwise it returns an object of class `mgNetworksCollection` which is a collection (a
-list) of `mgNetwork` objects. Below, we exemplify how to use the search functions, how to get a
-collection of networks and how to use other packages to carry out specific analyses.
+Note that if there is only one network to be retrieved, `get_collection()`
+returns a `mgNetwork` object, otherwise it returns an object of class
+`mgNetworksCollection` which is a collection (a list) of `mgNetwork` objects.
+Below, we exemplify how to use the search functions, how to get a collection of
+networks and how to use other packages to carry out specific analyses.
 
 ## Search functions
 
@@ -166,7 +167,6 @@ the corresponding table:
 
 
 Let's load **rmangal** as well as two helper packages:
-
 
 ```{R load_packages}
 library(rmangal)
@@ -229,23 +229,6 @@ insect_coll <- search_networks(query="insect%")
 glimpse(insect_coll)
 ```
 
-`search_networks()` allows spatial queries. The user can provide a
-[`sf`](https://cran.r-project.org/web/packages/sf/sf.pdf) object to access all
-networks included in the spatial extent of this object. For instance, we can request all
-networks in California:
-
-
-```{R}
-library(sf)
-library(mapview)
-library(USAboundaries)
-area <- us_states(state = "california")
-in_CA <- search_networks_sf(area)
-
-mapview(in_CA, legend = FALSE) + mapview(sf::st_cast(area,"MULTILINESTRING"), color = "red", legend = FALSE)
-```
-
-
 It is also possible to retrieve all networks based on interaction types involved:
 
 
@@ -255,6 +238,23 @@ avail_type()
 comp_interac <- search_interactions(type="competition")
 # Number of competition interactions in mangal
 nrow(comp_interac)
+```
+
+
+`search_networks_sf()` handles spatial queries: argument `query_sf` takes a
+[`sf`](https://cran.r-project.org/package=sf) object as input and returns all
+networks included in the spatial extent of this object. For instance, one can
+retrieve all Californian networks included in Mangal like so:
+
+
+```{R}
+library(sf)
+library(mapview)
+library(USAboundaries)
+area <- us_states(state = "california")
+in_CA <- search_networks_sf(area)
+
[…]
