### 2019-09-04 cbd470f
use devel in Appveyor :construction_worker:
