+++ appveyor.yml
@@ -9,6 +9,7 @@ install:
 
 environment:
   USE_RTOOLS: true
+  R_VERSION: devel
 
 # Adapt as necessary starting from here
 build_script:
