+++ appveyor.yml
@@ -8,9 +8,10 @@ install:
   ps: Bootstrap
 
 environment:
+  R_ARCH: x64
   USE_RTOOLS: true
   R_VERSION: devel
-  PKGTYPE: binary
+  PKGTYPE: source
 
 # Adapt as necessary starting from here
 build_script:
