### 2019-09-04 e33299d
additionnal tests
