+++ tests/testthat/test-search_interactions.R
@@ -4,6 +4,7 @@ res1 <- search_interactions(type = "competition")
 res2 <- search_interactions(type = "competition", expand_node = TRUE)
 res3 <- search_interactions(list(network_id = 926))
 res4 <- search_interactions(list(network_id = 926), expand_node = TRUE)
+
 test_that("expected behavior", {
   expect_identical(search_interactions("wrong"), data.frame())
   expect_equal(dim(res1), c(12, 19))
@@ -15,6 +16,8 @@ test_that("expected behavior", {
   expect_true(all(res3$network_id == 926))
   expect_equal(dim(res4), c(34, 57))
   expect_true(all(res4$network_id == 926))
+  expect_error(search_interactions(list(wrong = "wrong")))
+  expect_warning(search_interactions(list(network_id = 926, wrong = "wrong")))
 })
 
 
@@ -23,4 +26,5 @@ test_that("get_collection", {
   expect_equal(class(resc), "mgNetworksCollection")
   expect_equal(class(resc[[1L]]), "mgNetwork")
   expect_equal(names(resc[[1L]]), nm_co)
+
 })
