+++ appveyor.yml
@@ -11,7 +11,6 @@ environment:
   R_ARCH: x64
   USE_RTOOLS: true
   R_VERSION: devel
-  PKGTYPE: source
 
 # Adapt as necessary starting from here
 build_script:
