### 2019-09-05 #332 @SteveViss
We would like to thank again both reviewers for their constructive comments. Below, we first highlight the major changes we made in the package before addressing all comments individually.

NB: currently Appveyor is failing because binary for `ggraph 2.0.0` are not available yet.

## Mangal database

Some of the comments made by one of the reviewer suggested changes to the database. We considered and applied some changes but refrained to incorporate others that implied major changes on the data model and consequently trigger downstream changes on the API and the website. However, we kept track of these recommendations and may integrate these recommendations in future version of the infrastructure. Speaking of the future of Mangal, and to answer one comment, opening the database to publication is now our priority. We are currently thinking about a publication / submission process of new ecological networks using the `mgNetwork` class and a robust suite of tests to ensure data quality, and reduce curation time.

### Compute apriori networks properties

For the reason explained above, we did not prepopulate metrics (such as degree, centrality, connectance etc.) within the networks datatable. This will require extensive developments (for few functionalities) on the infrastructure backend. We have to change the model definition (i.e. add one by metric virtual / auto-generated field, compute metrics sometimes based on 2 other datatables: nodes and interactions).

We are also concerned about computing too many network descriptors / properties because several are sensitive to the algorithm selected. We basically recommend the user to do his own analysis through `igraph` (for instance), which is well-suited for this purpose. We however decided to provide a `summary` method that print a very simple set of network properties for `mgNetwork` objects (see comment [7872b50](https://github.com/mangal-wg/rmangal/commit/7872b50))

## Package features

### summary method

As mentioned above we added a summary methods for `mgNetwork` objects.

### mg_to_igraph becomes as.igraph

As suggested by one reviewer, instead of using `mg_to_igraph()` we extended the `as.igraph()` method for `mgNetwork` objects (see [1d47483](https://github.com/mangal-wg/rmangal/commit/1d47483)).  This makes the conversion to `igraph` and `tbl_graph` objects seamless.

### sf dependencies

`sf` has been moved to `Suggests`. Moreover, `sf` features are only used in `search_networks_sf()` and when argument `as_sf` is set to `TRUE`, see [#89](https://github.com/mangal-wg/rmangal/pull/89).

### search_references()

`search_references()` has been rewritten see [#85](https://github.com/mangal-wg/rmangal/pull/85);

## Documentation

### README

Now the README includes much more details about rmangal to provide a better description of the goals of rmangal and to conform with ROpenSci packaging guidelines, see commits [275f30a](https://github.com/mangal-wg/rmangal/commit/275f30a), [0ec0789](https://github.com/mangal-wg/rmangal/commit/0ec0789), [b1c0d83](https://github.com/mangal-wg/rmangal/commit/b1c0d83).

### API

We enhanced the API documentation by providing definition and type of all fields for each datatable. We made an explicit link to this in each R functions (see @references sections, [be50421](https://github.com/mangal-wg/rmangal/commit/be50421)).

### Functions documentation

We moved some text from the details section to the description as suggested by one reviewer.

### Vignette

We now provide an example of network manipulation with `tidygraph` and a an example of visualization with `ggraph`. See [46e4a63](https://github.com/mangal-wg/rmangal/commit/46e4a63). We also expand on how to use `sf` with rmangal (see [#89])

`search_references()` has been rewritten [#85];
`as_sf` is set to `TRUE` ;

### Website

The website has been rebuilt and it is now deployed by Travis CI. We also broke down the reference page to four categories of functions (https://mangal-wg.github.io/rmangal/reference/index.html).

## Tests

New features have all been covered in the unit testing and we added a little bit of icing on the cake!

## Review by @thomasp85

### Website

> One issue that could be adressed is the
> reference index where a bit of care could be used to group functions and give
> each group a short description. This will help new users a lot when navigating
> the website.

This has been done in [https://mangal.io/doc/r/reference/index.html](https://mangal.io/doc/r/reference/index.html)

### Dependencies

> The hard dependency on sf could be lifted as it doesn't appear to be used
> outside of a single functionality and sf has some very strong system
> dependencies. I suggest moving sf to suggest and testing for its availability
> when needed along with an error message about installing it if missing.

We have considered the recommandations (see [this issue](https://github.com/mangal-wg/rmangal/issues/83), and [0264bec](https://github.com/mangal-wg/rmangal/commit/0264bec)).

### Check

> R CMD check reports a range of examples with run time > 5sec and a total runtime
> of examples of 53 sec. This is likely to be flagged by CRAN during submission
> and it might as well be fixed beforehand:
>
>    checking examples (53.4s)
>    Examples with CPU or elapsed time > 5s
>                         user system elapsed
>    get_collection      4.812  0.069  22.876
>    combine_mgNetworks  1.129  0.056   6.831
>    search_interactions 0.718  0.010   6.440
>    search_datasets     0.369  0.008   5.359

We have now address this issue in commits [f31417c](https://github.com/mangal-wg/rmangal/commit/f31417c), [621dcb8](https://github.com/mangal-wg/rmangal/commit/621dcb8), [ebc4cf0](https://github.com/mangal-wg/rmangal/commit/ebc4cf0), [75dee5a](https://github.com/mangal-wg/rmangal/commit/75dee5a) by retrieving smaller networks and removing some examples. In improving examples, I noticed that examples are likely to have inconsistent timelapse because API responses are bandwidth dependent. If some issues are flagged by CRAN, we will use `\donttest`.

### Readme

> The README is a tad short. It does include installation instructions and some
> very useful links, but there is no motivating text for why this package may be
> of interest and what it tries to solve (besides being an API wrapper for a
> service). It should at least contain descriptions of why one might want to use
> the mangal database and an example of a network retrieval to show the strength
> of the API

This comment has been addressed above (see section *Readme*).

### Vignette

> This is a very good introduction to the package and shows of all important
> features in an easy to understand way. I would personally include a short
> section on visualising the networks, and maybe drag in tidygraph and ggraph
> (disclaimer: Author of both) for a more familiar interface to network analysis
> and visualisation than igraph (at least to the tidyverse crowd). I will not
> insist on this, though as package preferences are subjective Documentation

See section *Vignette* above.

> A few of the smaller functions lack examples, such a avail_type() and
> clear_cache_rmangal(). These are both functions without arguments so it is not
> clear that any example would add to the understanding, but in the interest of
> completeness  I'll just mention this.

`avail_type()` enumerate the list of interactions accessible from the database. `clear_cache_rmangal()` reset the memoise cache wrapped around `get_gen()` & `get_singleton()`. Both functions commit actions that do not required arguments. Because of that, we don't think any examples will improve functions usability.

> Most of the documentations have very short, sub-title
> like descriptions, with most of the text relegated to the Details section. This
> generally result in documentation entries where one have to read quite a bit
> before the function is justified and the experience could be improved by having
> a proper paragraph in the Description section.

We have improved functions description in [4271bf8](https://github.com/mangal-wg/rmangal/commit/4271bf8).

> Cross referencing could be improved. All functions that work with mgNetwork or
> mgNetworkCollection objects should link to or describe how one might obtain such
> an object. The same is true for functions working with mgSearch* objects.

This has been done in [0ec0789](https://github.com/mangal-wg/rmangal/commit/0ec0789)

> The
> key search_*() could also benefit from better crosslinking by using the
> '@family keyword from roxygen2. There are documentation entries for all methods
> to generics (pointing to the same doc). This makes the documentation index a bit
> messy so if this could be solved it would add to the quality.

> Small issues:
> combine_mgNetworks has a stray#'in the...` argument

This typo error has been fixed.

### Functionality

> The igraph integration is nice. With a very little modification you can make it
> work directly with tidygraph and ggraph.
>
> Either provide an explicit as.igraph method for mgNetwork object, which
> could simply be an alias for mg_to_igraph. Tidygraph and ggraph will look
> for such a method if no other exist and use that.
> Provide an as_tbl_graph method directly. This can again just be a wrapper
> around mg_to_igraph, e.g.
>
> as_tbl_graph.mgNetwork <- function(x, ...) {
>   as_tbl_graph(mg_to_igraph(x))
> }
>
> The obvious benefit of 1) is that you wouldn't need to list tidygraph as a
> dependency, while conversely the benefit of 2) is that you make support
> explicit

This comment has been addressed (see section *Package features*).

### Unit testing

> The package has a very good unit test coverage on all aspects of the code. The
> only area lacking is testing error handling as unit tests only tend to cover
> whether functions behave as expected with proper input. Tests for whether
> functions fails as expected and fails gracefully in the face of problems or bad
> input would be icing on the cake.

Addressed in section *Tests*.

## Review by @arw36

> It would be easy to prepopulate standard metrics. For the network datatable
> this would include # of nodes and interactions, potentially a modularity factor,
> and the interactions database with degree and eigen.centrality.

This comment has been answered in section *Compute apriori networks properties*.

> Unclear what the `df_net_unknown <- search_interactions(list(attr_id = 100),
> verbose = FALSE)` example is meant to accomplish in the search_interactions help
> file. This returned an empty dataframe for me.

We removed this example.

> May want to add `get_collection()` to the mg_insect example for
> `search_networks` documentation

Done!

> Suggest you add California border to the plot(networks_in_area) example

Done!

> I would like to see more metadata on variables including definitions for
> `attribute.id` and associated columns (name, description, unit) as well as
> `user_id`.

This has been answered in *Documentation* section.

> For the date in the edges table, it is unclear when this is extracted as a
> network-level trait, is an trait specific to each edge, or is NA.

We removed the field `date` in dataset table to avoid redundancies. We kept the `date` in the network and interaction datatables. In most cases, dates documented in edges and networks should be identical. But the network / sampling location could be visited several times (e.g. research would like to cover the flowering period). In that case, we keep track of the observation period through the field `date` in interaction datatable: the network datatable will contain the median date of the period and the interaction datable every observation dates.

> For instance,
> for edges of network_id = 19 it appears to be extrapolated from the network
> trait, despite the dataset_id (1) giving a range of dates (1897-09 to 1899-07).

Regarding the dataset #1, all informations (dataset, network and interactions) point to the same date: 1899-07-01. We are not able to reproduce the range of dates given by the reviewer (see code below).

```{r}
library(rmangal)
ds_1 <- search_datasets(list(id = 1))
net_19 <- get_collection(search_networks(list(id = 19)))
range(c(net_19$interactions$date, ds_1$date, net_19$network$date))
```

> Each data table starts with an `id` column, it may be worth clarifying:
> `network_id`, `edge_id`, `node_id`, e.t.c. in the main table. This will allow
> easy linking across tables, as ids are referenced by this in other tables (e.g.
> the `network_id` column in the nodes table).

We applied this suggestion on `mgNetwork` objects (see [c6cd9c0](https://github.com/mangal-wg/rmangal/commit/c6cd9c0)). However, the SQL side hasn't changed as it allows to make the distinction between primary key and foreign key.

> I suggest you make all your functions plural (currently `search_reference`
> is the only singular one.

`search_reference` is now renamed as `search_references` and `mgSearchReference`
as `mgSearchReferences` (see a98475b).

> I suggest you rename the author column in the reference table to be
> `first_author`

This suggestion has been taken into consideration.

> I was able to retrieve complete data tables for network, dataset, node,
> interactions using `search_*(“”)`. This did not work for the reference table.
> Ease of full database download would be nice, perhaps with warnings about
> download times and file sizes.

`search_references("")` is now working (see [cc297c84](https://github.com/mangal-wg/rmangal/commit/cc297c84))

> I think there should be greater curation/spell checks done in the
> interaction table, method column. What does biblio mean? Standardize
> capitalization.

We standardized capitalization and replaced `biblio` with `literature`.

> I find the categorization of interaction types to be non-standard. For
> instance, I would standardize null, NA, and none of interactions$attribute.unit
> to be the same (or clarify NA vs. none).

`null`, `NA` and `none` are now true SQL NULL.

> I think there is some confusion between edges/interactions. I wish the use of
> interaction was standardized in output text.

We replaced `edges` for `interactions` (see
[cd1430d](https://github.com/mangal-wg/rmangal/commit/cd1430d)) in all code.

> Currently, you can only `search_taxonomy` by the identified taxon rank. It
> would be really nice for broader rank taxon searches at the order or phylum
> level to retrieve all entries, even those that have a smaller scale taxonomy.
> For instance, as of now search_taxonomy(“Arthropoda”) returns no entries,
> despite there being lots of arthropods in the datasets.

Addressed in section *Mangal database* above.

> To conform with ROpenSci packaging guidelines, you will need to add brief
> demonstration usage, a description of how similar packages (i.e. `rglobi`)
> relate to this one, and citation information.

Usage, description and `rglobi` comparison have been added to the README except for the citation which can be found at https://mangal-wg.github.io/rmangal/authors.html or with `citation('rmangal')`

> In general, the ReadMe is short. Keep in mind this ReadMe may be first point
> of entry, so could be fleshed out a bit more with material you already have in
> vignettes/website.

See section *Readme* above.

> Do you expect users to upload networks to this database like `rglobi`? May
> be worth saying something about contributing new data or identifying new papers.

This comment has been discussed in section *Mangal database* above.
