### 2019-09-05 #332 @thomasp85
All my concerns have been addressed, but I will note (and urge a fix is made), that testing for the existence of `sf` should be done like:

```r
if (!requireNamespace("sf", quietly = TRUE)) {
  stop("sf should be installed to use this functionality", call. = FALSE) #or whatever wording you prefer
}
```

instead of the current call to `installed.packages()`
