### 2019-09-05 mangal-wg/rmangal#91 @KevCaz
Add reviewers in DESCRIPTION
Both reviewers agreed to do so and were very helpful, so let's add them!

https://github.com/ropensci/software-review/issues/332
