### 2019-09-05 mangal-wg/rmangal#92 @KevCaz
Properly test the existence of sf with requireNamespace
See comment by @thomasp85 https://github.com/ropensci/software-review/issues/332
