### 2019-09-05 mangal-wg/rmangal#93 @KevCaz
use spelling
See comment by @noamross in #90
and https://github.com/ropensci/software-review/issues/332
