+++ DESCRIPTION
@@ -16,8 +16,12 @@ Authors@R: c(
     person(given = "Dominique", family = "Gravel",
       email = "dominique.gravel@usherbrooke.ca", role = c("aut")),
     person(given = "Timothée", family = "Poisot",
-      email = "timothee.poisot@umontreal.ca", role = c("aut"))
-      )
+      email = "timothee.poisot@umontreal.ca", role = c("aut")),
+    person(given = "Thomas Lin", family = "Pedersen", role = "rev",
+           comment = c(ORCID = "0000-0002-5147-4711")),
+    person(given = "Anna", role = "rev",
+      comment = c(ORCID = "0000-0002-0504-0605"))
+    )
 Description: An interface to the Mangal database - a collection of ecological networks. This package includes functions to work with the Mangal RESTFul API methods (<https://mangal.io/doc/api>).
 URL: https://mangal.io, https://github.com/mangal-wg/rmangal
 BugReports: https://github.com/mangal-wg/rmangal/issues
