### 2019-09-05 fc9d798
add spell checking :sparkles: close #93
