+++ vignettes/rmangal.Rmd
@@ -90,7 +90,7 @@ specific description of the network.
 
 
 **Nodes** -- Information on the population, taxa or individu in the network. Each node has the
-original taxon name documented and taxonomic backbone provided by all services embbeded in taxize
+original taxon name documented and taxonomic backbone provided by all services embedded in taxize
 [@chamberlain_2019].
 
 **Interactions** -- Information on the interaction type (e.g. mutualism,
@@ -103,14 +103,14 @@ two nodes.
 
 
 
-## Authentification
+## Authentication
 
 
 So far, the `rmangal` package provides methods to get access to the data store. Data requests
-(performed via `httr::GET()`) do not require any authentification.
+(performed via `httr::GET()`) do not require any authentication.
 
 
-A bearer authentification strategy using [ORCID](https://orcid.org/) credentials
+A bearer authentication strategy using [ORCID](https://orcid.org/) credentials
 (as a third-party services) has been implemented on all `POST`, `DELETE`, `PUT`
 API operations to allow the user to add and delete new ecological to the data
 base. These features are not currently included in the **rmangal** package, but
@@ -352,7 +352,7 @@ mgn
 
 ## Geolocalize Mangal networks with `sf`
 
-The function `get_collection()` has an argument `as_sf` than converts network metadata of mgNetwork objacts to `sf` objects, which requires [`sf`](https://cran.r-project.org/package=sf) to be installed. This allows the user to easily
+The function `get_collection()` has an argument `as_sf` than converts network metadata of mgNetwork objects to `sf` objects, which requires [`sf`](https://cran.r-project.org/package=sf) to be installed. This allows the user to easily
 geolocalize the networks retrieved from Mangal.
 
 ```{R as_sf}
