+++ DESCRIPTION
@@ -41,6 +41,7 @@ Suggests:
     mapview,
     rmarkdown,
     sf,
+    spelling,
     taxize,
     tidygraph,
     testthat,
