### 2019-09-08 #332 @noamross
Thank you @SteveViss et al, @thomasp85 and @arw36, for an excellent review proccess! **rmangal is approved**.  🎉

To-dos:

- [ ] Add the review badge to your  README:

- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do and you can add any team members you wish.
- [ ] Update any links in badges for CI and coverage to point to the ropensci URL, and package.  Also point documentation urls to `https://docs.ropensci.org/{package}`. Docs should be built within 24 hours after transfer.

If/when you submit to a combined paper to JOSS, make a note in the pre-review thread that the R client has already been reviewed, linking to this thread, and tag `@ropensci/editors`. That way JOSS editors can limit review to the Julia client.

Our developer guide has a [section](https://ropensci.github.io/dev_guide/collaboration.html) for ongoing maintenance after acceptance. Welcome aboard!

This client but also the combined mangal.io ecosystem would be a great topic for a post on our blog.  If you are interested in writing something up please let @stefaniebutland know here. The timing is flexible can work with the other pieces of the system you are putting in place
