### 2019-09-09 #332 @SteveViss
We're grateful to the @ropensci team for this fantastic and constructive experience. Thank you @noamross, @thomasp85 & @arw36. We are doing the final checklist.

@arw36, we fully agree with your last comment. On the short term, we are tracking all data issues in https://github.com/mangal-wg/mangal-api/issues. Github is well suited for developers but not for all network ecologists. So on the long term, we may consider opening a [discourse](https://www.discourse.org/) platform which will be more inclusive and hopefully encourage users to flag potential data issues.

Hi @stefaniebutland ! We're interested in posting to the blog. [Mangal.io](https://mangal.io/#/) will be presented during the (BiodiversityNext)[https://biodiversitynext.org/] poster session (end of october). If possible, we propose to submit the pull request 2 weeks before that event (same for the JOSS paper).
