### 2019-09-09 1d7173c
Add ROpenSci badge
