+++ README.md
@@ -2,6 +2,7 @@
 
 # rmangal :package: - an R Client for the Mangal database
 
+[![](https://badges.ropensci.org/332_status.svg)](https://github.com/ropensci/software-review/issues/332)
 [![Travis-CI Build Status](https://travis-ci.org/mangal-wg/rmangal.svg?branch=master)](https://travis-ci.org/mangal-wg/rmangal)
 [![Build status](https://ci.appveyor.com/api/projects/status/mibs2ni969xiqgrd?svg=true)](https://ci.appveyor.com/project/KevCaz/rmangal)
 [![codecov](https://codecov.io/gh/mangal-wg/rmangal/branch/master/graph/badge.svg)](https://codecov.io/gh/mangal-wg/rmangal)
