+++ README.md
@@ -128,4 +128,11 @@ environmental gradients, which wasn't possible using the GloBi infrastructure.
 
 ## Code of conduct
 
-Please note that the 'rmangal' project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+Please note that the `rmangal` project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+
+## Meta
+
+* Get citation information for `rmangal` in R doing `citation(package = 'rmangal')`
+* Please [report any issues or bugs](https://github.com/ropensci/rmangal/issues).
+
+[![rofooter](https://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
