### 2019-09-09 mangal-wg/rmangal#91 @SteveViss
Done in cae8f017a8c0f3e15f8bf7017b9c5e5fda2c1373
