### 2019-09-09 #332 @SteveViss
Tues Oct 22 for the main conf. So our deadline should be Tue Oct 15, right? Let me know if you need more time.

> We can also coordinate this if you like: If you tweet to invite folks to your poster #, rOpenSci can quote tweet with a link to your post for more info and to bring more attention to your poster.

It will be great. Thank @stefaniebutland !
