+++ README.md
@@ -3,9 +3,9 @@
 # rmangal :package: - an R Client for the Mangal database
 
 [![](https://badges.ropensci.org/332_status.svg)](https://github.com/ropensci/software-review/issues/332)
-[![Travis-CI Build Status](https://travis-ci.org/mangal-wg/rmangal.svg?branch=master)](https://travis-ci.org/mangal-wg/rmangal)
+[![Build Status](https://travis-ci.org/ropensci/rmangal.svg?branch=master)](https://travis-ci.org/ropensci/rmangal)
 [![Build status](https://ci.appveyor.com/api/projects/status/mibs2ni969xiqgrd?svg=true)](https://ci.appveyor.com/project/KevCaz/rmangal)
-[![codecov](https://codecov.io/gh/mangal-wg/rmangal/branch/master/graph/badge.svg)](https://codecov.io/gh/mangal-wg/rmangal)
+[![codecov](https://codecov.io/gh/ropensci/rmangal/branch/master/graph/badge.svg)](https://codecov.io/gh/ropensci/rmangal)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![CRAN status](https://www.r-pkg.org/badges/version/rmangal)](http://www.r-pkg.org/badges/version/rmangal)
 
