### 2019-09-10 8bea1b9
prepare CRAN submission
