+++ cran-comments.md
@@ -1,0 +1,22 @@
+## This is the first submission
+
+This is version 2.0.0 because there was a version 1.x.x available on GitHub but there is no backward compatibility.
+
+## Test environments
+
+* local Debian testing (4.18.0-2-amd64), R-3.6.1
+* Ubuntu 14.04 (on TravisCI): R-release
+* Ubuntu 16.04 (on Travis-CI): R-devel
+* MacOSX 10.13.3 (on TravisCI), R-release
+* Windows (x86_64-w64-mingw32/x64, Windows Server 2012 om AppVeyor) R-release
+* Windows win-builder (R-release and R-devel)
+
+
+## R CMD check results
+
+0 ERRORs | 0 WARNINGs | 0 NOTES.
+
+
+## Downstream dependencies
+
+There are currently no downstream dependencies for this package.
