+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rmangal
 Type: Package
 Title: Mangal Client
-Version: 1.9.1.9000
+Version: 2.0.0
 Authors@R: c(
     person(given = "Steve", family = "Vissault", comment = c(ORCID = "0000-0002-0866-4376"),
       email = "steve.vissault@usherbrooke.ca", role = c("aut", "cre")),
@@ -19,12 +19,12 @@ Authors@R: c(
       email = "timothee.poisot@umontreal.ca", role = c("aut")),
     person(given = "Thomas Lin", family = "Pedersen", role = "rev",
            comment = c(ORCID = "0000-0002-5147-4711")),
-    person(given = "Anna", role = "rev",
+    person(given = "Anna Willoughby", role = "rev",
       comment = c(ORCID = "0000-0002-0504-0605"))
     )
 Description: An interface to the Mangal database - a collection of ecological networks. This package includes functions to work with the Mangal RESTFul API methods (<https://mangal.io/doc/api>).
-URL: https://mangal.io, https://github.com/mangal-wg/rmangal
-BugReports: https://github.com/mangal-wg/rmangal/issues
+URL: https://mangal.io, https://github.com/ropensci/rmangal
+BugReports: https://github.com/mangal-wg/ropensci/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
 Imports:
