### 2019-09-10 306780e
Fix URLs - CRAN presub
