+++ inst/bib/main.bib
@@ -9,8 +9,7 @@
   author = {Poisot, Timoth{\'e}e and Baiser, Benjamin and Dunne, Jennifer A. and K{\'e}fi, Sonia and Massol, Fran{\c c}ois and Mouquet, Nicolas and Romanuk, Tamara N. and Stouffer, Daniel B. and Wood, Spencer A. and Gravel, Dominique},
   month = apr,
   year = {2016},
-  pages = {384-390},
-  file = {/home/kevcaz/Zotero/storage/JJPGIQDM/Poisot et al._2016_mangal - making ecological network analysis simple.pdf}
+  pages = {384-390}
 }
 @Manual{chamberlain_2019,
   title = {taxize: Taxonomic information from around the web},
