### 2019-09-10 61d94da
type / minor changes for CRAN release
