+++ DESCRIPTION
@@ -22,7 +22,7 @@ Authors@R: c(
     person(given = "Anna Willoughby", role = "rev",
       comment = c(ORCID = "0000-0002-0504-0605"))
     )
-Description: An interface to the Mangal database - a collection of ecological networks. This package includes functions to work with the Mangal RESTFul API methods (<https://mangal.io/doc/api>).
+Description: An interface to the Mangal database - a collection of ecological networks. This package includes functions to work with the Mangal RESTful API methods (<https://mangal.io/doc/api>).
 URL: https://mangal.io, https://github.com/ropensci/rmangal
 BugReports: https://github.com/ropensci/rmangal/issues
 License: MIT + file LICENSE
