### 2019-09-11 c535df3
Fix URLs issues on README - CRAN presub
