### 2019-09-27 26229d1
CRAN submission
