+++ R/zzz.R
@@ -163,10 +163,6 @@ stop_if_missing_sf <- function(pkg = "sf") {
 #'
 #' @param endpoint `character` API entry point
 #' @param ... foreign key column name with the id
-#' @examples
-#'\dontrun{
-#' get_from_fkey(endpoints()$node, network_id = 926)
-#'}
 #' @return
 #' Object returned by [rmangal::get_gen()]
 #' @details
