### 2019-09-27 a97447a
Update man page
