### 2019-09-27 af3cb65
Add \donttest in search_interactions
