+++ R/search_interactions.R
@@ -37,7 +37,9 @@
 #' @examples
 #' df_inter <- search_interactions(type = "competition", verbose = FALSE)
 #' # Get all networks containing competition
+#' \donttest{
 #' competition_networks <- get_collection(df_inter, verbose = FALSE)
+#' }
 #' df_net_926 <- search_interactions(list(network_id = 926), verbose = FALSE)
 #' @export
 
