+++ man/search_interactions.Rd
@@ -49,7 +49,9 @@ Examples covering custom queries are provided below.
 \examples{
 df_inter <- search_interactions(type = "competition", verbose = FALSE)
 # Get all networks containing competition
+\donttest{
 competition_networks <- get_collection(df_inter, verbose = FALSE)
+}
 df_net_926 <- search_interactions(list(network_id = 926), verbose = FALSE)
 }
 \references{
