### 2019-09-27 d125ae5
Fix h1 and hex logo
