### 2019-10-09 #332 @KevCaz
@stefaniebutland I will be the one leading the blog post, we should be able to write something by Tuesday !
