### 2019-10-16 2d8872b
:bug: fix in print method for mgNetwork - close #94
