+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: rmangal
 Type: Package
 Title: 'Mangal' Client
-Version: 2.0.0
+Version: 2.0.1
 Authors@R: c(
     person(given = "Steve", family = "Vissault", comment = c(ORCID = "0000-0002-0866-4376"),
       email = "steve.vissault@usherbrooke.ca", role = c("aut", "cre")),
